added basic authentication

+++ controllers/households.js
@@ -1,16 +1,25 @@
 const householdsRouter = require('express').Router();
 const Household = require('../models/household');
 
+
 householdsRouter.get('/', async (request, response, next) => {
     try {
+        // Check authorization
+        if (!request.authorized) {
+            response.status(401).send({
+                error: "Invalid client key or secret"
+            });
+            return;
+        }
+
         Household.find({}).then(households => {
-            if (households) {
-                response.json(households);
-            } else {
-                response.status(404).end();
-            }
-        })
-        .catch(error => next(error))
+                if (households) {
+                    response.json(households);
+                } else {
+                    response.status(404).end();
+                }
+            })
+            .catch(error => next(error))
     } catch (error) {
         console.log('error', error);
         response.status(500).send({
@@ -22,14 +31,22 @@ householdsRouter.get('/', async (request, response, next) => {
 
 householdsRouter.get('/:id', async (request, response, next) => {
     try {
+        // Check authorization
+        if (!request.authorized) {
+            response.status(401).send({
+                error: "Invalid client key or secret"
+            });
+            return;
+        }
+
         Household.findById(request.params.id).then(household => {
-            if (household) {
-                response.json(household);
-            } else {
-                response.status(404).end();
-            }
-        })
-        .catch(error => next(error))
+                if (household) {
+                    response.json(household);
+                } else {
+                    response.status(404).end();
+                }
+            })
+            .catch(error => next(error))
     } catch (error) {
         console.log('error', error);
         response.status(500).send({
@@ -41,6 +58,14 @@ householdsRouter.get('/:id', async (request, response, next) => {
 
 householdsRouter.post('/', async (request, response, next) => {
     try {
+        // Check authorization
+        if (!request.authorized) {
+            response.status(401).send({
+                error: "Invalid client key or secret"
+            });
+            return;
+        }
+
         const body = request.body;
 
         if (body.name === undefined) {
@@ -55,9 +80,9 @@ householdsRouter.post('/', async (request, response, next) => {
         })
 
         household.save().then(savedHousehold => {
-            response.json(savedHousehold)
-        })
-        .catch(error => next(error))
+                response.json(savedHousehold)
+            })
+            .catch(error => next(error))
     } catch (error) {
         console.log('error', error);
         response.status(500).send({
@@ -69,6 +94,14 @@ householdsRouter.post('/', async (request, response, next) => {
 
 householdsRouter.delete('/:id', async (request, response, next) => {
     try {
+        // Check authorization
+        if (!request.authorized) {
+            response.status(401).send({
+                error: "Invalid client key or secret"
+            });
+            return;
+        }
+
         Household.findByIdAndRemove(request.params.id)
             .then(result => {
                 response.status(204).end()
@@ -84,6 +117,14 @@ householdsRouter.delete('/:id', async (request, response, next) => {
 
 householdsRouter.put('/:id', async (request, response, next) => {
     try {
+        // Check authorization
+        if (!request.authorized) {
+            response.status(401).send({
+                error: "Invalid client key or secret"
+            });
+            return;
+        }
+
         const body = request.body
 
         const household = {
